*: move "version" to "internal/version"
