+++ etcdctl/ctlv3/command/version_command.go
@@ -17,7 +17,8 @@ package command
 import (
 	"fmt"
 
-	"github.com/coreos/etcd/version"
+	"github.com/coreos/etcd/internal/version"
+
 	"github.com/spf13/cobra"
 )
