+++ etcdctl/ctlv3/help.go
@@ -25,7 +25,8 @@ import (
 	"text/tabwriter"
 	"text/template"
 
-	"github.com/coreos/etcd/version"
+	"github.com/coreos/etcd/internal/version"
+
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 )
