+++ etcdserver/api/capability.go
@@ -17,7 +17,8 @@ package api
 import (
 	"sync"
 
-	"github.com/coreos/etcd/version"
+	"github.com/coreos/etcd/internal/version"
+
 	"github.com/coreos/go-semver/semver"
 	"github.com/coreos/pkg/capnslog"
 )
