+++ etcdserver/api/etcdhttp/base.go
@@ -25,8 +25,9 @@ import (
 	"github.com/coreos/etcd/etcdserver"
 	"github.com/coreos/etcd/etcdserver/api"
 	"github.com/coreos/etcd/etcdserver/api/v2http/httptypes"
+	"github.com/coreos/etcd/internal/version"
 	"github.com/coreos/etcd/pkg/logutil"
-	"github.com/coreos/etcd/version"
+
 	"github.com/coreos/pkg/capnslog"
 )
