+++ etcdserver/api/etcdhttp/version_test.go
@@ -20,7 +20,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/coreos/etcd/version"
+	"github.com/coreos/etcd/internal/version"
 )
 
 func TestServeVersion(t *testing.T) {
