+++ etcdserver/cluster_util.go
@@ -23,8 +23,9 @@ import (
 	"time"
 
 	"github.com/coreos/etcd/etcdserver/membership"
+	"github.com/coreos/etcd/internal/version"
 	"github.com/coreos/etcd/pkg/types"
-	"github.com/coreos/etcd/version"
+
 	"github.com/coreos/go-semver/semver"
 )
