+++ etcdserver/cluster_util_test.go
@@ -18,8 +18,9 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/coreos/etcd/internal/version"
 	"github.com/coreos/etcd/pkg/types"
-	"github.com/coreos/etcd/version"
+
 	"github.com/coreos/go-semver/semver"
 )
