+++ etcdserver/membership/cluster.go
@@ -29,11 +29,11 @@ import (
 
 	"github.com/coreos/etcd/internal/mvcc/backend"
 	"github.com/coreos/etcd/internal/store"
+	"github.com/coreos/etcd/internal/version"
 	"github.com/coreos/etcd/pkg/netutil"
 	"github.com/coreos/etcd/pkg/types"
 	"github.com/coreos/etcd/raft"
 	"github.com/coreos/etcd/raft/raftpb"
-	"github.com/coreos/etcd/version"
 
 	"github.com/coreos/go-semver/semver"
 )
