+++ etcdserver/metrics.go
@@ -17,8 +17,9 @@ package etcdserver
 import (
 	"time"
 
+	"github.com/coreos/etcd/internal/version"
 	"github.com/coreos/etcd/pkg/runtime"
-	"github.com/coreos/etcd/version"
+
 	"github.com/prometheus/client_golang/prometheus"
 )
