+++ rafthttp/http.go
@@ -24,10 +24,11 @@ import (
 	"strings"
 
 	"github.com/coreos/etcd/internal/raftsnap"
+	"github.com/coreos/etcd/internal/version"
 	pioutil "github.com/coreos/etcd/pkg/ioutil"
 	"github.com/coreos/etcd/pkg/types"
+
 	"github.com/coreos/etcd/raft/raftpb"
-	"github.com/coreos/etcd/version"
 )
 
 const (
