+++ rafthttp/http_test.go
@@ -27,10 +27,10 @@ import (
 	"time"
 
 	"github.com/coreos/etcd/internal/raftsnap"
+	"github.com/coreos/etcd/internal/version"
 	"github.com/coreos/etcd/pkg/pbutil"
 	"github.com/coreos/etcd/pkg/types"
 	"github.com/coreos/etcd/raft/raftpb"
-	"github.com/coreos/etcd/version"
 )
 
 func TestServeRaftPrefix(t *testing.T) {
