+++ rafthttp/stream.go
@@ -28,11 +28,12 @@ import (
 	"golang.org/x/time/rate"
 
 	"github.com/coreos/etcd/etcdserver/stats"
+	"github.com/coreos/etcd/internal/version"
 	"github.com/coreos/etcd/pkg/httputil"
 	"github.com/coreos/etcd/pkg/transport"
 	"github.com/coreos/etcd/pkg/types"
 	"github.com/coreos/etcd/raft/raftpb"
-	"github.com/coreos/etcd/version"
+
 	"github.com/coreos/go-semver/semver"
 )
