+++ rafthttp/stream_test.go
@@ -26,14 +26,14 @@ import (
 	"testing"
 	"time"
 
-	"golang.org/x/time/rate"
-
 	"github.com/coreos/etcd/etcdserver/stats"
+	"github.com/coreos/etcd/internal/version"
 	"github.com/coreos/etcd/pkg/testutil"
 	"github.com/coreos/etcd/pkg/types"
 	"github.com/coreos/etcd/raft/raftpb"
-	"github.com/coreos/etcd/version"
+
 	"github.com/coreos/go-semver/semver"
+	"golang.org/x/time/rate"
 )
 
 // TestStreamWriterAttachOutgoingConn tests that outgoingConn can be attached
