+++ rafthttp/util_test.go
@@ -22,8 +22,9 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/coreos/etcd/internal/version"
 	"github.com/coreos/etcd/raft/raftpb"
-	"github.com/coreos/etcd/version"
+
 	"github.com/coreos/go-semver/semver"
 )
